+++ apps/consumer-client/.eslintrc.js
@@ -1,4 +1,4 @@
 module.exports = {
   extends: ["@pcd/eslint-config-custom"],
-  root: true,
+  root: true
 };
