+++ packages/lib/gpc/src/gpcTypes.ts
@@ -442,3 +442,22 @@ export type GPCRevealedClaims = {
    */
   watermark?: PODValue;
 };
+
+/**
+ * Converts a record of membership lists to one of membership sets.
+ *
+ * @param membershipLists the lists to convert
+ * @returns a record of membership sets
+ */
+export function membershipListsToSets(
+  membershipLists: PODMembershipLists
+): Record<PODName, Set<PODValue> | Set<PODValueTuple>> {
+  return Object.fromEntries(
+    Object.entries(membershipLists).map((pair) => [
+      pair[0],
+      new Set(pair[1] as (PODValue | PODValueTuple)[]) as
+        | Set<PODValue>
+        | Set<PODValueTuple>
+    ])
+  );
+}
